requirements.txt
fastapi
uvicorn[standard]
python-multipart
torch
torchvision
torchaudio
pillow
opencv-python
numpy
mediapipe
scikit-image
# dlib (optional; requires CMake; landmarks use MediaPipe/IrisLandmarkService)
google-genai
python-dotenv
pydantic-settings
protobuf
google-cloud-storage
motor
